packages/api/runtm_api/routes/health.py
"""Health check endpoint."""

from __future__ import annotations

from fastapi import APIRouter
from pydantic import BaseModel

router = APIRouter(tags=["health"])


class HealthResponse(BaseModel):
    """Health check response."""

    status: str
    version: str


@router.get("/health", response_model=HealthResponse)
async def health_check() -> HealthResponse:
    """Health check endpoint.

    Returns 200 OK if the service is healthy.
    """
    from runtm_api import __version__

    return HealthResponse(
        status="healthy",
        version=__version__,
    )
