packages/api/pyproject.toml
[project]
name = "runtm-api"
version = "0.1.0"
description = "FastAPI control plane for Runtm"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.9"
authors = [
    { name = "Gustavo Trigos", email = "gus@runtm.com" },
]
maintainers = [
    { name = "Gustavo Trigos", email = "gus@runtm.com" },
]
keywords = ["api", "fastapi", "deployment", "control-plane"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "runtm-shared>=0.1.0",
    "fastapi>=0.100.0,<1.0",
    "uvicorn[standard]>=0.20.0,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.10,<2.0",
    "asyncpg>=0.27,<1.0",
    "psycopg2-binary>=2.9,<3.0",
    "redis>=4.5,<6.0",
    "rq>=1.15,<2.0",
    "python-multipart>=0.0.6,<1.0",
    "pydantic-settings>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "pytest-asyncio>=0.21,<1.0",
    "httpx>=0.24,<1.0",
    "ruff>=0.1.0",
    "mypy>=1.0,<2.0",
]

[project.urls]
Homepage = "https://runtm.com"
Documentation = "https://docs.runtm.com"
Repository = "https://github.com/runtm-ai/runtm"
Issues = "https://github.com/runtm-ai/runtm/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["runtm_api"]

